Review comments

+++ src/hotspot/share/jvmci/jvmciCompilerToVM.cpp
@@ -1049,8 +1049,8 @@ C2V_END
 C2V_VMENTRY_0(jlong, getMaxCallTargetOffset, (JNIEnv* env, jobject, jlong addr))
   address target_addr = (address) addr;
   if (target_addr != 0x0) {
-    jlong off_low = (jlong)target_addr - ((jlong)CodeCache::low_bound() + sizeof(int));
-    jlong off_high = (jlong)target_addr - ((jlong)CodeCache::high_bound() + sizeof(int));
+    jlong off_low = (jlong)(target_addr - (CodeCache::low_bound() + sizeof(int)));
+    jlong off_high = (jlong)(target_addr - (CodeCache::high_bound() + sizeof(int)));
     return checked_cast<jlong>(MAX2(uabs(off_low), uabs(off_high)));
   }
   return -1;
